requirements.txt
# =============================================================================
# requirements.txt — Python Dependencies
# =============================================================================
# Install with:   pip install -r requirements.txt
#
# Pinned to minimum compatible versions.  Use `pip freeze > requirements.lock`
# for a fully reproducible lock file in CI/CD.
# =============================================================================

# ── Web Framework ────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0       # ASGI server with auto-reload & HTTP/2

# ── Settings / Environment ───────────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0       # BaseSettings for .env loading

# ── AI / Computer Vision ─────────────────────────────────────────────────────
ultralytics>=8.2.0              # YOLOv8 model loading & inference
opencv-python-headless>=4.9.0   # Image decoding (no GUI needed on servers)
numpy>=1.26.0
easyocr>=1.7.1                  # License plate text recognition (OCR)

# ── File Upload ──────────────────────────────────────────────────────────────
python-multipart>=0.0.9         # Required by FastAPI for file uploads

# ── Training & Export ────────────────────────────────────────────────────────
pyyaml>=6.0                     # YAML config parsing
onnx>=1.16.0                    # ONNX model format support
onnxruntime>=1.17.0             # ONNX inference runtime (CPU)
onnxslim>=0.1.0                 # ONNX graph simplification

# ── Future (uncomment when needed) ───────────────────────────────────────────
# easyocr>=1.7.1                # OCR — Step 2
# asyncpg>=0.29.0               # PostgreSQL async driver
# sqlalchemy[asyncio]>=2.0.0    # ORM
# alembic>=1.13.0               # DB migrations
